backend/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.28.0
fal-client>=0.5.0
